Adding new `ItemStore`, removed `MarkerStore` and `RouteStore`.
Since most things will share many properties and only differ on how
they are rendered, I moved the render specification into a field in a
JSON file. This disadvantage of this approach is that any large
changes will need to be effected across the entire file, so I may want
to add a script to do that for me.

I'm moving some of the view-specific state information outside of the
store; while storing that information was fast, it would be better to
track actions such as hovering inside only the view until the
information is needed on a global scale. It's not very MVC, but it
should save on performance for now.