Added `selected` state to each item, and responses for markers.
This is a bit weird, because markers will react to the event of
selecting a marker by centering the map on them and changing their
state. One alternative would be to make all state changes happen in
the ItemStore method `select`, and then have map actions happen in
another state change method.