Adding a stub for the Google Maps API.
Since the Google Maps API is not a node library, I don't think it will
be feasible to load it during testing without a lot of extra work. To
get around this, I'm faking API calls to only what I am using, and
then will check to see if the React components behave correctly in the
tests.