React components can be added as menu items using the `addControl` method.
The React menu controls are created separately from the Map `render()`
method. I think it is possible to have it in `render()`, but there
would be a lot of implementation-specific code that would take more
work and could break with later versions. The controls will assumably
not change their properties after creation, but we may want to check
this later on. It should be possible to re-render the components by
calling `ReactDOM.render()` again.