Added search functionality.
The item store now breaks names down by space, so that parts of the
word can be searched in addition to the whole name. Tries have been
modified to support multiple items at one branch to support the former
functionality.

Added an `AppStateStore` for managing the application state. The views
then check the application state and filter accordingly. Currently,
the only application states supported are "normal" and "search", but
more will likely be added in the future.