Adding offsets for routes.
Currently is not particularly complex; this depends on the fact that
there will only be three routes on the map. If we add more routes in
the future, we should probably look at dynamically changing route
offset based on how individual routes intersect.

The solution to this problem would be to compute where routes
intersect, and create a graph where nodes are routes and edges exist
when two routes intersect at some point. Then the problem of finding
the best offsets is equivalent to a graph coloring problem for
arbitrary graphs. This problem is generally NP-complete, so we should
probably assign the offsets using a separate program.