Added markers for photo opportunities, bike racks and emergency phones.
Even though this marker type is a little more limited, I'm setting up
the marker functionality so that it can adapt to missing data (since a
lot of features might be missing for future markers).

I added a new phone icon from www.mapbox.com/maki-icons/ that uses the
TU background for now. Same design as the markers before it.