Speeding up things again by having SearchResults cache the items it creates.
Whenever the component re-renders, it also must re-render every item
it creates. If this is done via a loop, then we have a problem. To
solve this issue, I create the items initially and then (plan to)
update it if we receive more items in the future.

This improves performance dramatically.